Replaced LRUCache with org.apache.commons.collections4.map.LRUMap

+++ src/main/java/GameShell.java
@@ -99,7 +99,7 @@ public abstract class GameShell extends Applet implements Runnable, MouseListene
 					ratio = lastRatio;
 					delta = lastDelta;
 				} else if (ntime > otim[opos]) {
-					ratio = (int) ((long) (2560L * deltime) / (ntime - otim[opos]));
+					ratio = (int) ((2560L * deltime) / (ntime - otim[opos]));
 				}
 
 				if (ratio < 25) {
